Extract ECS monitoring data collection to ServiceViewCollector
