+++ tests/unit/customizations/ecs/test_monitorexpressgatewayservice.py
@@ -123,617 +123,6 @@ class TestECSExpressGatewayServiceWatcher:
             ECSExpressGatewayServiceWatcher.is_monitoring_available() is False
         )
 
-    def setup_method(self):
-        self.app_session = create_app_session(output=DummyOutput())
-        self.app_session.__enter__()
-        self.mock_client = Mock()
-        self.service_arn = (
-            "arn:aws:ecs:us-west-2:123456789012:service/my-cluster/my-service"
-        )
-
-    def teardown_method(self):
-        if hasattr(self, 'app_session'):
-            self.app_session.__exit__(None, None, None)
-
-    def _create_watcher_with_mocks(self, resource_view="RESOURCE", timeout=1):
-        """Helper to create watcher with mocked display"""
-        mock_display = Mock()
-        mock_display.has_terminal.return_value = True
-        mock_display._check_keypress.return_value = None
-        mock_display._restore_terminal.return_value = None
-        mock_display.display.return_value = None
-
-        watcher = ECSExpressGatewayServiceWatcher(
-            self.mock_client,
-            self.service_arn,
-            resource_view,
-            timeout_minutes=timeout,
-            display=mock_display,
-        )
-
-        # Mock exec to call the monitoring method once and print output
-        original_monitor = watcher._monitor_express_gateway_service
-
-        def mock_exec():
-            try:
-                output = original_monitor("⠋", self.service_arn, resource_view)
-                print(output)
-                print("Monitoring Complete!")
-            except Exception as e:
-                # Re-raise expected exceptions
-                if isinstance(e, (ClientError, MonitoringError)):
-                    raise
-                # For other exceptions, just print and complete
-                print("Monitoring Complete!")
-
-        watcher.exec = mock_exec
-        return watcher
-
-    @patch('time.sleep')
-    def test_exec_successful_all_mode_monitoring(self, mock_sleep, capsys):
-        """Test successful monitoring in RESOURCE mode with resource parsing"""
-        watcher = self._create_watcher_with_mocks()
-        mock_sleep.side_effect = KeyboardInterrupt()
-
-        self.mock_client.describe_express_gateway_service.return_value = {
-            "service": {
-                "serviceArn": self.service_arn,
-                "cluster": "my-cluster",
-                "activeConfigurations": [{"serviceRevisionArn": "rev-arn"}],
-            }
-        }
-        self.mock_client.describe_service_revisions.return_value = {
-            "serviceRevisions": [
-                {
-                    "arn": "rev-arn",
-                    "ecsManagedResources": {
-                        "ingressPaths": [
-                            {
-                                "endpoint": "https://api.example.com",
-                                "loadBalancer": {
-                                    "arn": "arn:aws:elasticloadbalancing:us-west-2:123456789012:loadbalancer/app/my-lb/1234567890abcdef",
-                                    "status": "ACTIVE",
-                                },
-                                "targetGroups": [
-                                    {
-                                        "arn": "arn:aws:elasticloadbalancing:us-west-2:123456789012:targetgroup/my-tg/1234567890abcdef",
-                                        "status": "HEALTHY",
-                                    }
-                                ],
-                            }
-                        ],
-                        "serviceSecurityGroups": [
-                            {
-                                "arn": "arn:aws:ec2:us-west-2:123456789012:security-group/sg-1234567890abcdef0",
-                                "status": "ACTIVE",
-                            }
-                        ],
-                        "logGroups": [
-                            {
-                                "arn": "arn:aws:logs:us-west-2:123456789012:log-group:/aws/ecs/my-service",
-                                "status": "ACTIVE",
-                            }
-                        ],
-                    },
-                }
-            ]
-        }
-        self.mock_client.describe_services.return_value = {
-            "services": [{"events": [{"message": "Running"}]}]
-        }
-
-        watcher.exec()
-        captured = capsys.readouterr()
-        output_text = captured.out
-
-        # Verify parsed resources appear in output
-        assert "Cluster" in output_text
-        assert "Service" in output_text
-        assert "IngressPath" in output_text
-        assert "LoadBalancer" in output_text
-        assert "TargetGroup" in output_text
-        assert "SecurityGroup" in output_text
-        assert "LogGroup" in output_text
-
-        # Specific identifiers
-        assert "https://api.example.com" in output_text  # IngressPath endpoint
-        assert "my-lb" in output_text  # LoadBalancer identifier
-        assert "my-tg" in output_text  # TargetGroup identifier
-        assert (
-            "sg-1234567890abcdef0" in output_text
-        )  # SecurityGroup identifier
-        assert "/aws/ecs/my-service" in output_text  # LogGroup identifier
-
-        # Status values
-        assert "ACTIVE" in output_text  # LoadBalancer and SecurityGroup status
-        assert "HEALTHY" in output_text  # TargetGroup status
-
-    @patch('time.sleep')
-    def test_exec_successful_delta_mode_with_deployment(
-        self, mock_sleep, capsys
-    ):
-        """Test DEPLOYMENT mode executes successfully"""
-        watcher = self._create_watcher_with_mocks()
-        mock_sleep.side_effect = KeyboardInterrupt()
-
-        self.mock_client.describe_express_gateway_service.return_value = {
-            "service": {
-                "serviceArn": self.service_arn,
-                "cluster": "my-cluster",
-                "activeConfigurations": [],
-            }
-        }
-        self.mock_client.describe_services.return_value = {
-            "services": [{"events": [{"message": "Service running"}]}]
-        }
-
-        watcher.exec()
-        captured = capsys.readouterr()
-
-        # Verify DEPLOYMENT mode executes successfully
-        assert captured.out
-
-    @patch('time.sleep')
-    def test_exec_combined_view_multiple_revisions(self, mock_sleep, capsys):
-        """Test RESOURCE mode combines multiple service revisions correctly"""
-        watcher = self._create_watcher_with_mocks()
-        mock_sleep.side_effect = KeyboardInterrupt()
-
-        # Multiple active configurations (combined view)
-        self.mock_client.describe_express_gateway_service.return_value = {
-            "service": {
-                "serviceArn": self.service_arn,
-                "cluster": "my-cluster",
-                "activeConfigurations": [
-                    {"serviceRevisionArn": "rev-1"},
-                    {"serviceRevisionArn": "rev-2"},
-                ],
-            }
-        }
-
-        # Mock multiple revisions with different resources
-        self.mock_client.describe_service_revisions.return_value = {
-            "serviceRevisions": [
-                {
-                    "arn": "rev-1",
-                    "ecsManagedResources": {
-                        "ingressPaths": [
-                            {
-                                "endpoint": "https://api.example.com",
-                                "loadBalancer": {
-                                    "arn": "arn:aws:elasticloadbalancing:us-west-2:123456789012:loadbalancer/app/api-lb/1234",
-                                    "status": "ACTIVE",
-                                },
-                            }
-                        ],
-                        "serviceSecurityGroups": [
-                            {
-                                "arn": "arn:aws:ec2:us-west-2:123456789012:security-group/sg-api123",
-                                "status": "ACTIVE",
-                            }
-                        ],
-                    },
-                },
-                {
-                    "arn": "rev-2",
-                    "ecsManagedResources": {
-                        "ingressPaths": [
-                            {
-                                "endpoint": "https://web.example.com",
-                                "loadBalancer": {
-                                    "arn": "arn:aws:elasticloadbalancing:us-west-2:123456789012:loadbalancer/app/web-lb/5678",
-                                    "status": "CREATING",
-                                },
-                            }
-                        ],
-                        "logGroups": [
-                            {
-                                "arn": "arn:aws:logs:us-west-2:123456789012:log-group:/aws/ecs/web-logs",
-                                "status": "ACTIVE",
-                            }
-                        ],
-                    },
-                },
-            ]
-        }
-
-        self.mock_client.describe_services.return_value = {
-            "services": [
-                {"events": [{"message": "Multiple revisions active"}]}
-            ]
-        }
-
-        watcher.exec()
-        captured = capsys.readouterr()
-        output_text = captured.out
-
-        # Verify combined view shows resources from both revisions
-        # Resource types from both revisions
-        assert "IngressPath" in output_text
-        assert "LoadBalancer" in output_text
-        assert "SecurityGroup" in output_text  # From rev-1
-        assert "LogGroup" in output_text  # From rev-2
-
-        # Specific identifiers from both revisions
-        assert "https://api.example.com" in output_text  # From rev-1
-        assert "https://web.example.com" in output_text  # From rev-2
-        assert "api-lb" in output_text  # From rev-1
-        assert "web-lb" in output_text  # From rev-2
-        assert "sg-api123" in output_text  # From rev-1
-        assert "/aws/ecs/web-logs" in output_text  # From rev-2
-
-        # Status values from both revisions
-        assert "ACTIVE" in output_text  # From both revisions
-        assert "CREATING" in output_text  # From rev-2
-
-    @patch('time.sleep')
-    def test_exec_keyboard_interrupt_handling(self, mock_sleep, capsys):
-        watcher = self._create_watcher_with_mocks()
-        mock_sleep.side_effect = KeyboardInterrupt()
-
-        self.mock_client.describe_express_gateway_service.return_value = {
-            "service": {
-                "serviceArn": self.service_arn,
-                "cluster": "my-cluster",
-                "activeConfigurations": [],
-            }
-        }
-
-        watcher.exec()
-        captured = capsys.readouterr()
-
-        # Verify completion message is printed
-        assert "Monitoring Complete!" in captured.out
-
-    @patch('time.sleep')
-    def test_exec_with_service_not_found_error(self, mock_sleep):
-        """Test exec() with service not found error bubbles up"""
-        watcher = self._create_watcher_with_mocks()
-        mock_sleep.side_effect = KeyboardInterrupt()
-
-        error = ClientError(
-            error_response={
-                'Error': {
-                    'Code': 'ServiceNotFoundException',
-                    'Message': 'Service not found',
-                }
-            },
-            operation_name='DescribeExpressGatewayService',
-        )
-        self.mock_client.describe_express_gateway_service.side_effect = error
-
-        with pytest.raises(ClientError) as exc_info:
-            watcher.exec()
-
-        # Verify the specific error is raised
-        assert (
-            exc_info.value.response['Error']['Code']
-            == 'ServiceNotFoundException'
-        )
-        assert (
-            exc_info.value.response['Error']['Message'] == 'Service not found'
-        )
-
-    @patch('time.sleep')
-    def test_exec_with_inactive_service_handled_gracefully(
-        self, mock_sleep, capsys
-    ):
-        """Test exec() handles inactive service gracefully"""
-        watcher = self._create_watcher_with_mocks()
-        mock_sleep.side_effect = KeyboardInterrupt()
-
-        self.mock_client.describe_express_gateway_service.side_effect = ClientError(
-            error_response={
-                'Error': {
-                    'Code': 'InvalidParameterException',
-                    'Message': 'Cannot call DescribeServiceRevisions for a service that is INACTIVE',
-                }
-            },
-            operation_name='DescribeExpressGatewayService',
-        )
-        self.mock_client.describe_services.return_value = {
-            "services": [{"events": [{"message": "Service is inactive"}]}]
-        }
-
-        watcher.exec()
-        captured = capsys.readouterr()
-
-        # Verify inactive service is handled and appropriate message shown
-        assert "inactive" in captured.out.lower()
-
-    @patch('time.sleep')
-    def test_exec_with_empty_resources(self, mock_sleep, capsys):
-        """Test parsing edge case: empty/null resources"""
-        watcher = self._create_watcher_with_mocks()
-        mock_sleep.side_effect = KeyboardInterrupt()
-
-        self.mock_client.describe_express_gateway_service.return_value = {
-            "service": {
-                "serviceArn": self.service_arn,
-                "cluster": "my-cluster",
-                "activeConfigurations": [{"serviceRevisionArn": "rev-arn"}],
-            }
-        }
-        # Empty ecsManagedResources
-        self.mock_client.describe_service_revisions.return_value = {
-            "serviceRevisions": [{"arn": "rev-arn", "ecsManagedResources": {}}]
-        }
-        self.mock_client.describe_services.return_value = {
-            "services": [{"events": [{"message": "No resources"}]}]
-        }
-
-        watcher.exec()
-        captured = capsys.readouterr()
-        output_text = captured.out
-
-        # Should handle empty resources gracefully but still show basic structure
-        assert "Cluster" in output_text
-        assert "Service" in output_text
-        # Should NOT contain resource types since ecsManagedResources is empty
-        assert "IngressPath" not in output_text
-        assert "LoadBalancer" not in output_text
-
-    @patch('time.sleep')
-    def test_exec_with_autoscaling_resources(self, mock_sleep, capsys):
-        """Test autoscaling resource parsing with scalableTarget and policies"""
-        watcher = self._create_watcher_with_mocks()
-        mock_sleep.side_effect = KeyboardInterrupt()
-
-        self.mock_client.describe_express_gateway_service.return_value = {
-            "service": {
-                "serviceArn": self.service_arn,
-                "cluster": "my-cluster",
-                "activeConfigurations": [{"serviceRevisionArn": "rev-arn"}],
-            }
-        }
-        self.mock_client.describe_service_revisions.return_value = {
-            "serviceRevisions": [
-                {
-                    "arn": "rev-arn",
-                    "ecsManagedResources": {
-                        "autoScaling": {
-                            "scalableTarget": {
-                                "arn": "arn:aws:application-autoscaling:us-west-2:123456789012:scalable-target/1234567890abcdef",
-                                "status": "ACTIVE",
-                            },
-                            "applicationAutoScalingPolicies": [
-                                {
-                                    "arn": "arn:aws:application-autoscaling:us-west-2:123456789012:scaling-policy/cpu-policy",
-                                    "status": "ACTIVE",
-                                },
-                                {
-                                    "arn": "arn:aws:application-autoscaling:us-west-2:123456789012:scaling-policy/memory-policy",
-                                    "status": "ACTIVE",
-                                },
-                            ],
-                        }
-                    },
-                }
-            ]
-        }
-        self.mock_client.describe_services.return_value = {
-            "services": [{"events": [{"message": "Autoscaling active"}]}]
-        }
-
-        watcher.exec()
-        captured = capsys.readouterr()
-        output_text = captured.out
-
-        assert "AutoScaling" in output_text
-        assert "ScalableTarget" in output_text
-        assert "AutoScalingPolicy" in output_text
-        # ScalableTarget identifier
-        assert "1234567890abcdef" in output_text
-        # Policy identifiers
-        assert "cpu-policy" in output_text
-        assert "memory-policy" in output_text
-
-    @patch('time.sleep')
-    def test_exec_with_malformed_resource_data(self, mock_sleep, capsys):
-        """Test parsing edge case: malformed resource data"""
-        watcher = self._create_watcher_with_mocks()
-        mock_sleep.side_effect = KeyboardInterrupt()
-
-        self.mock_client.describe_express_gateway_service.return_value = {
-            "service": {
-                "serviceArn": self.service_arn,
-                "cluster": "my-cluster",
-                "activeConfigurations": [{"serviceRevisionArn": "rev-arn"}],
-            }
-        }
-        # Malformed resources - missing required fields
-        self.mock_client.describe_service_revisions.return_value = {
-            "serviceRevisions": [
-                {
-                    "arn": "rev-arn",
-                    "ecsManagedResources": {
-                        "ingressPaths": [
-                            {"endpoint": "https://example.com"}
-                        ],  # Missing loadBalancer
-                        "serviceSecurityGroups": [
-                            {"status": "ACTIVE"}
-                        ],  # Missing arn
-                    },
-                }
-            ]
-        }
-        self.mock_client.describe_services.return_value = {
-            "services": [{"events": [{"message": "Malformed data"}]}]
-        }
-
-        watcher.exec()
-        captured = capsys.readouterr()
-        output_text = captured.out
-
-        # Should handle malformed data gracefully and show what it can parse
-        assert "IngressPath" in output_text
-        assert "https://example.com" in output_text
-        # Should show SecurityGroup type even with missing arn
-        assert "SecurityGroup" in output_text
-        # Should NOT show LoadBalancer since it's missing from IngressPath
-        assert "LoadBalancer" not in output_text
-
-    @patch('time.sleep')
-    def test_exec_eventually_consistent_missing_deployment(
-        self, mock_sleep, capsys
-    ):
-        """Test eventually consistent behavior: deployment missing after list"""
-        watcher = self._create_watcher_with_mocks()
-        mock_sleep.side_effect = KeyboardInterrupt()
-
-        self.mock_client.describe_express_gateway_service.return_value = {
-            "service": {
-                "serviceArn": self.service_arn,
-                "cluster": "my-cluster",
-                "activeConfigurations": [],
-            }
-        }
-        # List shows deployment exists
-        self.mock_client.list_service_deployments.return_value = {
-            "serviceDeployments": [{"serviceDeploymentArn": "deploy-arn"}]
-        }
-        # But describe fails (eventually consistent)
-        self.mock_client.describe_service_deployments.return_value = {
-            "serviceDeployments": [],
-            "failures": [{"arn": "deploy-arn", "reason": "MISSING"}],
-        }
-        self.mock_client.describe_services.return_value = {
-            "services": [{"events": [{"message": "Eventually consistent"}]}]
-        }
-
-        watcher.exec()
-        captured = capsys.readouterr()
-        output_text = captured.out
-
-        # Should handle eventually consistent missing deployment gracefully
-        # Should show waiting state when deployment is missing
-        assert "Trying to describe gateway service" in output_text
-        assert "Monitoring Complete" in output_text
-
-    @patch('time.sleep')
-    def test_exec_eventually_consistent_missing_revision(
-        self, mock_sleep, capsys
-    ):
-        """Test eventually consistent behavior: service revision missing after deployment describe"""
-        watcher = self._create_watcher_with_mocks()
-        mock_sleep.side_effect = KeyboardInterrupt()
-
-        self.mock_client.describe_express_gateway_service.return_value = {
-            "service": {
-                "serviceArn": self.service_arn,
-                "cluster": "my-cluster",
-                "activeConfigurations": [],
-            }
-        }
-        self.mock_client.list_service_deployments.return_value = {
-            "serviceDeployments": [{"serviceDeploymentArn": "deploy-arn"}]
-        }
-        self.mock_client.describe_service_deployments.return_value = {
-            "serviceDeployments": [
-                {
-                    "serviceDeploymentArn": "deploy-arn",
-                    "status": "IN_PROGRESS",
-                    "targetServiceRevision": {"arn": "target-rev"},
-                }
-            ]
-        }
-        # Service revision missing (eventually consistent)
-        self.mock_client.describe_service_revisions.return_value = {
-            "serviceRevisions": [],
-            "failures": [{"arn": "target-rev", "reason": "MISSING"}],
-        }
-        self.mock_client.describe_services.return_value = {
-            "services": [{"events": [{"message": "Revision missing"}]}]
-        }
-
-        watcher.exec()
-        captured = capsys.readouterr()
-        output_text = captured.out
-
-        # Should handle eventually consistent missing revision gracefully
-        # Should show waiting state when revision is missing
-        assert "Trying to describe gateway service" in output_text
-        assert "Monitoring Complete" in output_text
-
-    @patch('time.sleep')
-    def test_exec_with_api_failures(self, mock_sleep):
-        """Test failure parsing: API returns failures"""
-        watcher = self._create_watcher_with_mocks()
-        mock_sleep.side_effect = KeyboardInterrupt()
-
-        self.mock_client.describe_express_gateway_service.return_value = {
-            "service": {
-                "serviceArn": self.service_arn,
-                "cluster": "my-cluster",
-                "activeConfigurations": [{"serviceRevisionArn": "rev-arn"}],
-            }
-        }
-        # API returns failures
-        self.mock_client.describe_service_revisions.return_value = {
-            "serviceRevisions": [],
-            "failures": [{"arn": "rev-arn", "reason": "ServiceNotFound"}],
-        }
-
-        with pytest.raises(MonitoringError) as exc_info:
-            watcher.exec()
-
-        # Should raise MonitoringError with failure details
-        error_message = str(exc_info.value)
-        assert "DescribeServiceRevisions" in error_message
-        assert "rev-arn" in error_message
-        assert "ServiceNotFound" in error_message
-
-    @patch('time.sleep')
-    def test_exec_with_malformed_api_failures(self, mock_sleep):
-        """Test failure parsing: malformed failure responses"""
-        watcher = self._create_watcher_with_mocks()
-        mock_sleep.side_effect = KeyboardInterrupt()
-
-        self.mock_client.describe_express_gateway_service.return_value = {
-            "service": {
-                "serviceArn": self.service_arn,
-                "cluster": "my-cluster",
-                "activeConfigurations": [{"serviceRevisionArn": "rev-arn"}],
-            }
-        }
-        # Malformed failures - missing arn or reason
-        self.mock_client.describe_service_revisions.return_value = {
-            "serviceRevisions": [],
-            "failures": [{"reason": "ServiceNotFound"}],  # Missing arn
-        }
-
-        with pytest.raises(MonitoringError) as exc_info:
-            watcher.exec()
-
-        # Should raise MonitoringError about invalid failure response
-        error_message = str(exc_info.value)
-        assert "Invalid failure response" in error_message
-        assert "missing arn or reason" in error_message
-
-    @patch('time.sleep')
-    def test_exec_with_missing_response_fields(self, mock_sleep):
-        """Test response validation: missing required fields"""
-        watcher = self._create_watcher_with_mocks()
-        mock_sleep.side_effect = KeyboardInterrupt()
-
-        self.mock_client.describe_express_gateway_service.return_value = {
-            "service": {
-                "serviceArn": self.service_arn,
-                "cluster": "my-cluster",
-                "activeConfigurations": [{"serviceRevisionArn": "rev-arn"}],
-            }
-        }
-        # Missing serviceRevisions field
-        self.mock_client.describe_service_revisions.return_value = {}
-
-        with pytest.raises(MonitoringError) as exc_info:
-            watcher.exec()
-
-        # Should raise MonitoringError about empty response
-        error_message = str(exc_info.value)
-        assert "DescribeServiceRevisions" in error_message
-        assert "empty" in error_message
-
 
 class TestMonitoringError:
     """Test MonitoringError exception class"""
